Whole lotta updates - video.js was able to squeeze SSR fix out into v7.12.1 - Removed pinned version of videojs-vtt.js - Broken webpack config out into separate dev and prod files - Webpack configs both merge in a common config - Updated scripts to use respective webpack config - Webpack prod output auto assigns videojs
